Remove pose selector service try/catch block
... and add more comments.

The wait_for_service calls were moved out of the block in 6ff3da4a6.
Now, the whole block does not make sense any more, so it is removed.

Without the try/catch, wait_for_service or creating the ServiceProxys
will throw an uncaught ROSException, and the whole node exits (same
behavior as with the block).

We only need the special handling with the try/catch and explicit
signal_shutdown after we call moveit_commander.roscpp_initialize() later
on, because it will overwrite the signal handler, so the node will no
longer exit after an uncaught exception.