Fix table plane adjustment parameters
Before, the allowed placement area was the size of the table plane,
*extended* by 0.05 in the x direction (and no adjustment in the y
direction).

After this commit, it's the table plane, reduced by 0.15 m in x and y
direction (because the long side of the KLT box is around 0.30 m, so
with this adjustment it always fits fully onto the table).